refactor: extract attendance app from blowcomotion monolith
