+++ attendance/templates/attendance/capture.html
