+++ attendance/templates/attendance/capture_success.html
