+++ attendance/templates/attendance/partials/all_reports_content.html
