+++ attendance/templates/attendance/partials/capture_content.html
