+++ attendance/templates/attendance/partials/capture_success.html
