+++ attendance/templates/attendance/partials/member_entry.html
