+++ attendance/templates/attendance/partials/section_content.html
