+++ attendance/templates/attendance/section_report.html
