+++ attendance/tests/test_attendance_commands.py
